+++ favorites/models.py
@@ -14,6 +14,23 @@ class FavoriteManager(models.Manager):
         """
         return ContentType.objects.get_for_model(type(obj))
 
+    def get_favorite(self, obj, user):
+        """
+        Get the favorite 
+        """
+        if not user.is_authenticated():
+            return False
+            
+        content_type = ContentType.objects.get_for_model(type(obj))
+        return self.get_query_set().get(content_type=self.get_content_type(obj),
+            object_id=obj.pk, user=user)
+
+    def is_favorite(self, obj, user):
+        """
+        Check if its in your favs already
+        """
+        return 
+
     @classmethod
     def add_favorite(cls, user, content_object):
         """