+++ CHANGELOG.rst
@@ -3,6 +3,12 @@
 ChangeLog
 =========
 
+0.1.2
+------
+- Demo project context processors
+- Optional GET args handled in context processors
+- View-topic page
+
 0.1.1
 -----
 - Base init tests